find ID and PW function added & changed names of files
