+++ src/main/java/com/conan/semi/account/CheckNicknameC.java
@@ -11,7 +11,7 @@ import java.io.IOException;
 public class CheckNicknameC extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         // 닉네임 중복확인
-        String result = AccountDAO_test3.checkNicknameExists(request);
+        String result = AccountDAO.checkNicknameExists(request);
 
         // 응답을 텍스트 형식으로 클라이언트에게 반환
         response.setContentType("text/plain");
