+++ src/main/java/com/conan/semi/login/Find_IDC.java
@@ -10,9 +10,15 @@ import java.io.IOException;
 @WebServlet("/Find_IDC")
 public class Find_IDC extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-    request.getRequestDispatcher("jsp/login/login_find_id.jsp").forward(request, response);
+    request.getRequestDispatcher("jsp/login/login_find_id_input.jsp").forward(request, response);
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    request.setCharacterEncoding("utf-8");
+    // 아이디 찾기
+    FindDAO.findID(request);
+
+    // 어디로?
+    request.getRequestDispatcher("jsp/login/login_find_id_output.jsp").forward(request, response);
     }
 }
