+++ src/main/java/com/conan/semi/login/Find_PWC.java
@@ -10,9 +10,15 @@ import java.io.IOException;
 @WebServlet("/Find_PWC")
 public class Find_PWC extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-    request.getRequestDispatcher("jsp/login/login_find_pw.jsp").forward(request, response);
+    request.getRequestDispatcher("jsp/login/login_find_pw_input.jsp").forward(request, response);
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    request.setCharacterEncoding("utf-8");
+    // 비밀번호 찾기
+    FindDAO.findPW(request);
+
+    // 어디로?
+    request.getRequestDispatcher("jsp/login/login_find_pw_output.jsp").forward(request, response);
     }
 }
