+++ src/main/java/com/conan/semi/login/LoginC.java
@@ -10,12 +10,14 @@ import java.io.IOException;
 @WebServlet("/LoginC")
 public class LoginC extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.getRequestDispatcher("jsp/login/login.jsp").forward(request, response);
+        request.getRequestDispatcher("jsp/login/login_input.jsp").forward(request, response);
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
     // 로그인
     LoginDAO_test.login(request);
+
     // 어디로?
+    request.getRequestDispatcher("jsp/login/login_output.jsp").forward(request, response);
     }
 }
