--- src/main/java/com/conan/semi/login/LoginC_test.java
@@ -1,23 +1,0 @@
-package com.conan.semi.login;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet("/LoginC_test")
-public class LoginC_test extends HttpServlet {
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.getRequestDispatcher("jsp/login/login_test.jsp").forward(request, response);
-    }
-
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-    // 로그인
-    LoginDAO_test.login(request);
-
-    // 어디로?
-    request.getRequestDispatcher("jsp/main_test.jsp").forward(request, response);
-    }
-}
