+++ src/main/java/com/conan/semi/main/HC.java
@@ -11,8 +11,8 @@ import java.io.IOException;
 @WebServlet("/HC")
 public class HC extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-    request.getRequestDispatcher("jsp/main_test.jsp").forward(request, response);
-    } // used by hdh
+    request.getRequestDispatcher("jsp/main_HDH.jsp").forward(request, response);
+    } // used by HDH
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
     }
