+++ src/main/webapp/jsp/account/account_test2.jsp
@@ -374,7 +374,7 @@
 
     <div>
         <button type="submit" disabled> REGISTER</button> <!-- 초기 상태에서는 비활성화 -->
-        <button type="button" class="back-btn" onclick="location.href='jsp/main_test.jsp'">메인으로</button>
+        <button type="button" class="back-btn" onclick="location.href='jsp/main_HDH.jsp'">메인으로</button>
     </div>
 </form>
